[gonra] update entidades
