+++ api/src/main/java/com/atos/inventario/model/Documento.java
@@ -2,21 +2,42 @@ package com.atos.inventario.model;
 
 import java.util.Date;
 
-import javax.persistence.ManyToOne;
+import javax.persistence.*;
 
+@Entity
+@Inheritance(strategy = InheritanceType.TABLE_PER_CLASS)
 public abstract class Documento {
 	
-	@ManyToOne
+	@Id
+	@GeneratedValue(strategy = GenerationType.AUTO)
+	private long id;
+	
+	@ManyToOne(cascade = CascadeType.ALL)
+	@JoinColumn(name = "cod_unidade_produtora", referencedColumnName = "idUnidadeProdutora")
 	private UnidadeProdutora unidadeProdutora;
-	@ManyToOne
+	
+	@ManyToOne(cascade = CascadeType.ALL)
+	@JoinColumn(name = "cod_classificaco_documental", referencedColumnName = "codigoClassificacaoDocumental")
 	private ClassificacaoDocumental classificacaoDocumental;
+	
 	private Date dataLimite;
-	@ManyToOne
+	
+	@ManyToOne(cascade = CascadeType.ALL)
+	@JoinColumn(name = "cod_empregado", referencedColumnName = "idEmpregado")
 	private Empregado empregado;
+	
 	private Date dataCriacao;
-	@ManyToOne
+	
+	@ManyToOne(cascade = CascadeType.ALL)
+	@JoinColumn(name = "cod_localizacao", referencedColumnName = "idLocalizacao")
 	private Localizacao localizacao;
 	
+	public long getId() {
+		return id;
+	}
+	public void setId(long id) {
+		this.id = id;
+	}
 	public UnidadeProdutora getUnidadeProdutora() {
 		return unidadeProdutora;
 	}
@@ -53,6 +74,5 @@ public abstract class Documento {
 	public void setLocalizacao(Localizacao localizacao) {
 		this.localizacao = localizacao;
 	}
-	
 	
 }
