+++ api/src/main/java/com/atos/inventario/model/OutroDocumento.java
@@ -15,26 +15,11 @@ public class OutroDocumento extends Documento implements Serializable {
 	
 	private static final long serialVersionUID = 1L;
 
-	@Id
-	@GeneratedValue(strategy = GenerationType.AUTO)
-	private long id;
-
 	private String documentoEncaminhamento;
 	private String numeroCaixaEscritorioOrigem;
 	private String numeroCaixaArquivoCustodia;
 	private String objetoResumido;
-
 	
-	
-	
-	public long getId() {
-		return id;
-	}
-
-	public void setId(long id) {
-		this.id = id;
-	}
-
 	public String getDocumentoEncaminhamento() {
 		return documentoEncaminhamento;
 	}
