+++ api/src/main/java/com/atos/inventario/model/PastaFuncional.java
@@ -14,23 +14,9 @@ public class PastaFuncional extends Documento implements Serializable {
 
 	private static final long serialVersionUID = 1L;
 
-	@Id
-	@GeneratedValue(strategy = GenerationType.AUTO)
-	private long id;
-
 	private String documentoEncaminhamento;
 	private String numeroCaixaEscritorioOrigem;
 	private String numeroCaixaArquivoCustodia;
-	
-
-	
-	public long getId() {
-		return id;
-	}
-
-	public void setId(long id) {
-		this.id = id;
-	}
 
 	public String getDocumentoEncaminhamento() {
 		return documentoEncaminhamento;
